(#89) RequestedFields Should Use API Names - Uses reflection to validate fields to ensure they exist on    GlossaryTerm - Broke out some tests into partial classes for easier readability - Added in VS noise introduced that I am tiring of removing
