+++ src/NCI.OCPL.Api.Glossary/NCI.OCPL.Api.Glossary.csproj
@@ -13,7 +13,9 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.App" Version="2.1.4"/>
+    <PackageReference Include="Microsoft.AspNetCore.App" Version="2.1.4"><PrivateAssets Condition="'%(PackageReference.Version)' == ''">all</PrivateAssets>
+<Publish Condition="'%(PackageReference.Version)' == ''">true</Publish>
+</PackageReference>
     <PackageReference Include="Microsoft.AspNetCore.Razor.Design" Version="2.1.2" PrivateAssets="All" />
     <PackageReference Include="NSwag.AspNetCore" Version="11.17.1" />
     <PackageReference Include="NEST" Version="5.6.1" />
