+++ test/NCI.OCPL.Api.Glossary.Tests/Tests/Services/ESTermsQueryServiceTest.Expand.cs
@@ -1,166 +1,30 @@
 using System;
-using Elasticsearch.Net;
+using System.Collections.Generic;
+
+using Microsoft.Extensions.Logging.Testing;
 using Microsoft.Extensions.Options;
-using Moq;
+
+using Elasticsearch.Net;
+using Nest;
+using Newtonsoft.Json.Linq;
+using Xunit;
+
 using NCI.OCPL.Api.Common.Testing;
 using NCI.OCPL.Api.Glossary.Models;
 using NCI.OCPL.Api.Glossary.Services;
 using NCI.OCPL.Api.Glossary.Tests.ESTermsQueryTestData;
-using Nest;
-using Microsoft.Extensions.Logging.Testing;
-using System.Collections.Generic;
-using System.IO;
-using Xunit;
 using NCI.OCPL.Api.Common;
-using Newtonsoft.Json.Linq;
-using NCI.OCPL.Api.Glossary.Tests;
+
 
 namespace NCI.OCPL.Api.Glossary.Tests
 {
-    public class ESTermsQueryServiceTest
+    public partial class ESTermsQueryServiceTest
     {
-
-        public static IEnumerable<object[]> GetByIdData => new[] {
-            new object[] { new GetById_43966_NoMediaNoResources() },
-            new object[] { new GetById_44058_VideoExernalLink() },
-            new object[] { new GetById_44759_NoMediaGlossaryResource() },
-            new object[] { new GetById_44178_NoMediaSummaryLink() },
-            new object[] { new GetById_44386_NoMediaDrugSummary() },
-            new object[] { new GetById_339337_HealthProfessional() },
-            new object[] { new GetById_445043_ImageAndExternalLink() },
-        };
 
         public static IEnumerable<object[]> ExpandData => new[] {
             new object[] { new Expand_S() },
             new object[] { new Expand_NoResults() }
         };
-
-        #region GetById tests
-
-        /// <summary>
-        /// Test failure to connect to and retrieve response from API in GetById.
-        /// </summary>
-        [Fact]
-        public async void GetById_TestAPIConnectionFailure()
-        {
-            ElasticsearchInterceptingConnection conn = new ElasticsearchInterceptingConnection();
-            conn.RegisterRequestHandlerForType<Nest.GetResponse<GlossaryTerm>>((req, res) =>
-            {
-                res.StatusCode = 500;
-            });
-
-            // While this has a URI, it does not matter, an InMemoryConnection never requests
-            // from the server.
-            var pool = new SingleNodeConnectionPool(new Uri("http://localhost:9200"));
-
-            var connectionSettings = new ConnectionSettings(pool, conn);
-            IElasticClient client = new ElasticClient(connectionSettings);
-
-            // Setup the mocked Options
-            IOptions<GlossaryAPIOptions> gTermsClientOptions = GetMockOptions();
-
-            ESTermsQueryService termsClient = new ESTermsQueryService(client, gTermsClientOptions, new NullLogger<ESTermsQueryService>());
-
-            APIErrorException ex = await Assert.ThrowsAsync<APIErrorException>(() => termsClient.GetById("cancer.gov", AudienceType.Patient, "en", 43966L, new string[] { }));
-            Assert.Equal(500, ex.HttpStatusCode);
-        }
-
-        /// <summary>
-        /// Test failure to connect to ES or receiving an invalid response from ES in GetById.
-        /// </summary>
-        [Fact]
-        public async void GetById_TestInvalidResponse()
-        {
-            ElasticsearchInterceptingConnection conn = new ElasticsearchInterceptingConnection();
-            conn.RegisterRequestHandlerForType<Nest.GetResponse<GlossaryTerm>>((req, res) =>
-            {
-
-            });
-
-            // While this has a URI, it does not matter, an InMemoryConnection never requests
-            // from the server.
-            var pool = new SingleNodeConnectionPool(new Uri("http://localhost:9200"));
-
-            var connectionSettings = new ConnectionSettings(pool, conn);
-            IElasticClient client = new ElasticClient(connectionSettings);
-
-            // Setup the mocked Options
-            IOptions<GlossaryAPIOptions> gTermsClientOptions = GetMockOptions();
-
-            ESTermsQueryService termsClient = new ESTermsQueryService(client, gTermsClientOptions, new NullLogger<ESTermsQueryService>());
-
-            APIErrorException ex = await Assert.ThrowsAsync<APIErrorException>(() => termsClient.GetById("cancer.gov", AudienceType.Patient, "en", 43966L, new string[] { }));
-            Assert.Equal(500, ex.HttpStatusCode);
-        }
-
-        /// <summary>
-        /// Test that GetById URI for Elasticsearch is set up correctly.
-        /// </summary>
-        [Theory, MemberData(nameof(GetByIdData))]
-        public async void GetById_TestUriSetup(BaseTermsQueryTestData data)
-        {
-            Uri esURI = null;
-
-            ElasticsearchInterceptingConnection conn = new ElasticsearchInterceptingConnection();
-            conn.RegisterRequestHandlerForType<Nest.GetResponse<GlossaryTerm>>((req, res) =>
-            {
-                //Get the file name for this round
-                res.Stream = TestingTools.GetTestFileAsStream("ESTermsQueryData/GetById/" + data.ESTermID + ".json");
-
-                res.StatusCode = 200;
-
-                esURI = req.Uri;
-            });
-
-            // While this has a URI, it does not matter, an InMemoryConnection never requests
-            // from the server.
-            var pool = new SingleNodeConnectionPool(new Uri("http://localhost:9200"));
-
-            var connectionSettings = new ConnectionSettings(pool, conn);
-            IElasticClient client = new ElasticClient(connectionSettings);
-
-            // Setup the mocked Options
-            IOptions<GlossaryAPIOptions> gTermsClientOptions = GetMockOptions();
-
-            ESTermsQueryService termsClient = new ESTermsQueryService(client, gTermsClientOptions, new NullLogger<ESTermsQueryService>());
-
-            // We don't actually care that this returns anything - only that the intercepting connection
-            // sets up the request URI correctly.
-            GlossaryTerm actDisplay = await termsClient.GetById(
-                data.DictionaryName,
-                data.Audience,
-                data.Language,
-                data.TermID,
-                new string[] { }
-            );
-
-            Assert.Equal( $"/glossaryv1/terms/{data.ESTermID}", esURI.AbsolutePath);
-        }
-
-        /// <summary>
-        /// Tests the correct loading of various data files for GetById.
-        /// </summary>
-        /// <param name="data"></param>
-        /// <returns></returns>
-        [Theory, MemberData(nameof(GetByIdData))]
-        public async void GetById_DataLoading(BaseTermsQueryTestData data)
-        {
-            IElasticClient client = GetById_GetElasticClientWithData(data);
-
-            // Setup the mocked Options
-            IOptions<GlossaryAPIOptions> gTermsClientOptions = GetMockOptions();
-
-            ESTermsQueryService termsClient = new ESTermsQueryService(client, gTermsClientOptions, new NullLogger<ESTermsQueryService>());
-
-            GlossaryTerm glossaryTerm = await termsClient.GetById("cancer.gov", AudienceType.Patient, "en", 43966L, new string[] { });
-
-            Assert.Equal(data.ExpectedData, glossaryTerm, new GlossaryTermComparer());
-        }
-
-        #endregion
-
-
-        #region Expand tests
 
         /// <summary>
         /// Test failure to connect to and retrieve response from API in GetById
@@ -237,9 +101,9 @@ namespace NCI.OCPL.Api.Glossary.Tests
                             ""audience"",
                             ""term_name"",
                             ""first_letter"",
-                            ""pretty_url_name"",
+                            ""pretty_url_name"",                            
+                            ""pronunciation"",
                             ""definition"",
-                            ""pronunciation""
                         ]
                     },
                     ""sort"": [
@@ -308,7 +172,7 @@ namespace NCI.OCPL.Api.Glossary.Tests
 
             try
             {
-                var results = await termsClient.Expand("Cancer.gov", AudienceType.Patient, "en", "s", 5, 0, new string[]{"term_id", "language", "dictionary", "audience", "term_name", "first_letter", "pretty_url_name", "definition", "pronunciation"});
+                var results = await termsClient.Expand("Cancer.gov", AudienceType.Patient, "en", "s", 5, 0, new string[]{"termId", "language", "dictionary", "audience", "termName", "firstLetter", "prettyUrlName", "definition", "pronunciation"});
             }
             catch (Exception) { }
 
@@ -330,37 +194,53 @@ namespace NCI.OCPL.Api.Glossary.Tests
 
             ESTermsQueryService termsClient = new ESTermsQueryService(client, gTermsClientOptions, new NullLogger<ESTermsQueryService>());
 
-            GlossaryTermResults glossaryTermResults = await termsClient.Expand("Cancer.gov", AudienceType.Patient, "en", "s", 5, 0, new string[]{"term_id", "language", "dictionary", "audience", "term_name", "first_letter", "pretty_url_name", "definition", "pronunciation"});
+            GlossaryTermResults glossaryTermResults = await termsClient.Expand("Cancer.gov", AudienceType.Patient, "en", "s", 5, 0, new string[]{"termId", "language", "dictionary", "audience", "termName", "firstLetter", "prettyUrlName", "definition", "pronunciation"});
 
             Assert.Equal(data.ExpectedData.Results, glossaryTermResults.Results, new GlossaryTermComparer());
             Assert.Equal(data.ExpectedData.Meta.TotalResults, glossaryTermResults.Meta.TotalResults);
             Assert.Equal(data.ExpectedData.Meta.From, glossaryTermResults.Meta.From);
         }
 
-        #endregion
-
-        ///<summary>
-        ///A private method to enrich data from file for GetById
-        ///</summary>
-        private IElasticClient GetById_GetElasticClientWithData(BaseTermsQueryTestData data)
+        /// <summary>
+        /// Tests the correct loading of various data files.
+        /// </summary>
+        /// <param name="data"></param>
+        /// <returns></returns>
+        [Fact]
+        public async void Expand_ValidRequestedFields()
         {
             ElasticsearchInterceptingConnection conn = new ElasticsearchInterceptingConnection();
-            conn.RegisterRequestHandlerForType<Nest.GetResponse<GlossaryTerm>>((req, res) =>
+            conn.RegisterRequestHandlerForType<Nest.SearchResponse<GlossaryTerm>>((req, res) =>
             {
-                //Get the file name for this round
-                res.Stream = TestingTools.GetTestFileAsStream("ESTermsQueryData/GetById/" + data.ESTermID + ".json");
 
-                res.StatusCode = 200;
             });
 
-            //While this has a URI, it does not matter, an InMemoryConnection never requests
-            //from the server.
+            // While this has a URI, it does not matter, an InMemoryConnection never requests
+            // from the server.
             var pool = new SingleNodeConnectionPool(new Uri("http://localhost:9200"));
 
             var connectionSettings = new ConnectionSettings(pool, conn);
             IElasticClient client = new ElasticClient(connectionSettings);
 
-            return client;
+            // Setup the mocked Options
+            IOptions<GlossaryAPIOptions> gTermsClientOptions = GetMockOptions();
+
+            ESTermsQueryService termsClient = new ESTermsQueryService(client, gTermsClientOptions, new NullLogger<ESTermsQueryService>());
+
+            // Test null argument
+            await Assert.ThrowsAsync<ArgumentNullException>(async () => {
+                GlossaryTermResults glossaryTermResults = await termsClient.Expand("Cancer.gov", AudienceType.Patient, "en", "s", 5, 0, null);
+            });
+
+            // Test null in an array
+            await Assert.ThrowsAsync<ArgumentException>(async () => {
+                GlossaryTermResults glossaryTermResults = await termsClient.Expand("Cancer.gov", AudienceType.Patient, "en", "s", 5, 0, new string[] { null });
+            });
+
+            // Test bad property name
+            await Assert.ThrowsAsync<ArgumentOutOfRangeException>(async () => {
+                GlossaryTermResults glossaryTermResults = await termsClient.Expand("Cancer.gov", AudienceType.Patient, "en", "s", 5, 0, new string[] { "chicken" });
+            });
         }
 
         ///<summary>
@@ -387,21 +267,5 @@ namespace NCI.OCPL.Api.Glossary.Tests
             return client;
         }
 
-        ///<summary>
-        ///A private method to enrich IOptions
-        ///</summary>
-        private IOptions<GlossaryAPIOptions> GetMockOptions()
-        {
-            Mock<IOptions<GlossaryAPIOptions>> glossaryAPIClientOptions = new Mock<IOptions<GlossaryAPIOptions>>();
-            glossaryAPIClientOptions
-                .SetupGet(opt => opt.Value)
-                .Returns(new GlossaryAPIOptions()
-                {
-                    AliasName = "glossaryv1"
-                }
-            );
-
-            return glossaryAPIClientOptions.Object;
-        }
     }
 }
